Added Data Classes And Optimised the Graphql Query
