+++ data/src/main/java/com/practical/data/EpsoideDataMapper.kt
@@ -10,11 +10,8 @@ fun EpisodeQuery.Episode.toEpisodeModelDetails(): EpisodeModelDetails {
     return EpisodeModelDetails(
         data = EpisodeData(
             episode = Episode(
-                id = this.id,
                 name = this.name,
                 airDate = this.air_date,
-                episode = this.episode,
-                created = this.created,
                 characters = this.characters.mapNotNull { it?.toCharacter() }
             )
         )
@@ -24,13 +21,7 @@ fun EpisodeQuery.Episode.toEpisodeModelDetails(): EpisodeModelDetails {
 fun EpisodeQuery.Character.toCharacter(): Character {
     return Character(
         id = this.id,
-        name = this.name,
-        status = this.status,
-        species = this.species,
-        type = this.type,
-        gender = this.gender,
-        image = this.image,
-        created = this.created
+        image = this.image
     )
 }
