+++ domain/src/main/java/com/practical/domain/Character.kt
@@ -1,0 +1,7 @@
+package com.practical.domain
+
+data class Character(
+    val id: String,
+    val image: String,
+)
+
