+++ domain/src/main/java/com/practical/domain/Episode.kt
@@ -1,0 +1,7 @@
+package com.practical.domain
+
+data class Episode(
+    val airDate: String,
+    val characters: List<Character>,
+    val name: String,
+)
