+++ domain/src/main/java/com/practical/domain/EpisodeData.kt
@@ -1,0 +1,6 @@
+package com.practical.domain
+
+data class EpisodeData(
+    val episode: Episode,
+)
+
